Adapt to proxy# having different specificities across GHC versions
Prior to GHC 8.10, the kind variable in the type of `proxy#` was specified,
whereas in 8.10 or later, it it inferred. See
https://gitlab.haskell.org/ghc/ghc/-/issues/16293. As a result,
`proxy# @(a :: k)` is well kinded on 8.10 or later, but not on older GHCs.
Bummer.

This works around the issue by using `proxy# :: Proxy# (a :: k)` instead, which
should be compatible with all supported GHC versions.